+++ .eslintrc
@@ -28,6 +28,7 @@
     "no-mixed-spaces-and-tabs": ["error", "smart-tabs"],
     "no-multi-spaces": ["error", { "ignoreEOLComments": false }],
     "no-multiple-empty-lines": ["error", { "max": 1 }],
+    "no-underscore-dangle": "off",
     "no-unused-expressions": "off",
     "no-unused-vars": [
       "warn",
