+++ src/server/routes/items.js
@@ -1,0 +1,10 @@
+const express = require('express');
+
+const router = express.Router();
+
+const { getAllItems } = require('../controllers/items');
+
+// Get all items
+router.get('/:type', getAllItems);
+
+module.exports = router;
